{"schema":"decodex/commit/1","summary":"Modularize overlay session bootstrap","authority":"manual"}
